+++ HISTORY.md
@@ -6,6 +6,7 @@
 
 #### API
 
+- Add `replace` parameter to sample_elites in archives ({pr}`682`)
 - Require numpy to be at least 2.0.0 ({pr}`681`)
 - Allow specifying `centroids` with filenames in CVTArchive ({pr}`679`)
 - Add `kdtree_query_kwargs` parameter to CVTArchive ({pr}`677`)